# docs/feed-pulse/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Feed Pulse application.",
        "title": "Feed Pulse API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Login a user with username/email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate a user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "login successful response",
                        "schema": {
                            "$ref": "#/definitions/httpUtils.HTTPMessage"
                        }
                    },
                    "400": {
                        "description": "bad request error",
                        "schema": {
                            "$ref": "#/definitions/httpUtils.HTTPError"
                        }
                    },
                    "401": {
                        "description": "authentication error",
                        "schema": {
                            "$ref": "#/definitions/httpUtils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "user not found error",
                        "schema": {
                            "$ref": "#/definitions/httpUtils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpUtils.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginUser": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "httpUtils.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "bad request"
                }
            }
        },
        "httpUtils.HTTPMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}